Create README.md
Easier access to medical records has enabled various machine learning algorithms to achieve high success rates on early diagnosis and treatment of cancer. Globally, about 1 in 6 deaths is due to cancer in the world. Spotting cancer early increases the chances of survival. Diagnosing cancer before it has the chance to spread too far means that treatment is more likely to be successful. In this study lymphoma cancer and the subtypes which are Chronic lymphocytic, Mantle cell lymphoma, Follicular lymphoma, diagnosed in patients is combined with powerful mathematical algorithms. The goal is to perfectly analyse the data, with different algorithms such as Convolutional Neural Networks (CNNs) and CNN+LSTM mixed models to obtain high accuracy rates on classifying the lymphoma subtypes. These algorithms are applied on both RGB and grayscale images. In order to get the best result from each algorithm, appropriate data preprocessing methods and to avoid overfitting regularization techniques such as dropout was used. The algorithms are compared afterwards.